@@ -5,19 +5,23 @@ namespace PHPStamp\Document;
 use PHPStamp\Document\WordDocument\Cleanup;
 use PHPStamp\Exception\InvalidArgumentException;
 use PHPStamp\Extension\Extension;
+use PHPStamp\Extension\ExtensionInterface;
 use PHPStamp\Processor\Tag;
 
 /**
- * @link http://msdn.microsoft.com/ru-ru/library/office/gg278327(v=office.15).aspx
+ * @see http://msdn.microsoft.com/ru-ru/library/office/gg278327(v=office.15).aspx
  */
 class WordDocument extends Document
 {
-    private $structure = array('w:p', 'w:r', 'w:rPr', 'w:t');
+    /**
+     * @var array<string>
+     */
+    private array $structure = ['w:p', 'w:r', 'w:rPr', 'w:t'];
 
     /**
      * Path to main content file inside document ZIP archive.
      */
-    public function getContentPath()
+    public static function getContentPath()
     {
         return 'word/document.xml';
     }
@@ -25,9 +29,11 @@ class WordDocument extends Document
     /**
      * Get node name by XPATH_* constant type.
      *
-     * @param int $type XPATH_* constant.
-     * @param bool $global Append global xpath //.
+     * @param int  $type   XPATH_* constant
+     * @param bool $global append global xpath //
+     *
      * @return string
+     *
      * @throws InvalidArgumentException
      */
     public function getNodeName($type, $global = false)
@@ -36,7 +42,7 @@ class WordDocument extends Document
             throw new InvalidArgumentException('Element with this index not defined in structure');
         }
 
-        $return = array();
+        $return = [];
         if ($global === true) {
             $return[] = '//';
         }
@@ -56,7 +62,6 @@ class WordDocument extends Document
     /**
      * Cleanup Word Document from WYSIWYG mess.
      *
-     * @param \DOMDocument $template
      * @throws InvalidArgumentException
      */
     public function cleanup(\DOMDocument $template)
@@ -77,23 +82,26 @@ class WordDocument extends Document
     /**
      * Get instance of associated placeholder function.
      *
-     * @param string $id Id as entered in placeholder.
-     * @param Tag $tag Container tag.
+     * @param string $id  id as entered in placeholder
+     * @param Tag    $tag container tag
+     *
      * @return Extension
+     *
      * @throws InvalidArgumentException
      */
-    public function getExpression($id, Tag $tag)
+    public function getExpression(string $id, Tag $tag): ExtensionInterface
     {
-        $available = array(
-			'cell' => 'PHPStamp\\Document\\WordDocument\\Extension\\Cell',
-			'listitem' => 'PHPStamp\\Document\\WordDocument\\Extension\\ListItem',
-		);
-		
-		if (isset($available[$id]) === false) {
-			throw new InvalidArgumentException('Class by id "' . $id . '" not found.');
-		}
-		
+        $available = [
+            'cell' => 'PHPStamp\\Document\\WordDocument\\Extension\\Cell',
+            'listitem' => 'PHPStamp\\Document\\WordDocument\\Extension\\ListItem',
+        ];
+
+        if (isset($available[$id]) === false) {
+            throw new InvalidArgumentException('Class by id "'.$id.'" not found.');
+        }
+
         $className = $available[$id];
+
         return new $className($tag);
     }
 }
