@@ -6,10 +6,7 @@ use PHPStamp\Processor\Tag;
 
 abstract class Extension implements ExtensionInterface
 {
-    /**
-     * @var Tag
-     */
-    protected $tag;
+    protected Tag $tag;
 
     /**
      * @inherit
@@ -30,16 +27,19 @@ abstract class Extension implements ExtensionInterface
 
     /**
      * Prepare / validate / merge with defaults / modify given arguments.
-     * @param array $arguments
-     * @return array
+     *
+     * @param array<string> $arguments
+     *
+     * @return array<string>
      */
-    protected abstract function prepareArguments(array $arguments);
+    abstract protected function prepareArguments(array $arguments);
 
     /**
      * All template modification magic is here.
-     * @param array $arguments
-     * @param \DOMElement $node
+     *
+     * @param array<string> $arguments
+     *
      * @return void
      */
-    protected abstract function insertTemplateLogic(array $arguments, \DOMElement $node);
+    abstract protected function insertTemplateLogic(array $arguments, \DOMElement $node);
 }
