@@ -2,15 +2,17 @@
 
 namespace PHPStamp\Document;
 
+use PHPStamp\Exception\HashException;
 use PHPStamp\Exception\InvalidArgumentException;
+use PHPStamp\Extension\ExtensionInterface;
 use PHPStamp\Processor\Tag;
 
 abstract class Document implements DocumentInterface
 {
-    const XPATH_PARAGRAPH  = 0;
-    const XPATH_RUN  = 1;
-    const XPATH_RUN_PROPERTY  = 2;
-    const XPATH_TEXT  = 3;
+    public const XPATH_PARAGRAPH = 0;
+    public const XPATH_RUN = 1;
+    public const XPATH_RUN_PROPERTY = 2;
+    public const XPATH_TEXT = 3;
 
     /**
      * Original document filename.
@@ -30,6 +32,7 @@ abstract class Document implements DocumentInterface
      * Creates a new Document.
      *
      * @param string $documentPath
+     *
      * @throws InvalidArgumentException
      */
     public function __construct($documentPath)
@@ -45,26 +48,26 @@ abstract class Document implements DocumentInterface
     /**
      * Extract main content file.
      *
-     * @param string $to Path to extract content file.
-     * @param bool $overwrite Overwrite content file.
-     * @return string Full path to extracted document file.
+     * @param string $to        path to extract content file
+     * @param bool   $overwrite overwrite content file
+     *
+     * @return string full path to extracted document file
+     *
      * @throws InvalidArgumentException
      */
     public function extract($to, $overwrite)
     {
-        $filePath = $to . $this->getDocumentName() . '/' . $this->getContentPath();
+        $filePath = $to.$this->getDocumentName().'/'.$this->getContentPath();
 
         if (!file_exists($filePath) || $overwrite === true) {
             $zip = new \ZipArchive();
 
             $code = $zip->open($this->getDocumentPath());
             if ($code !== true) {
-                throw new InvalidArgumentException(
-                    'Can`t open archive "' . $this->documentPath . '", code "' . $code . '" returned.'
-                );
+                throw new InvalidArgumentException('Can`t open archive "'.$this->documentPath.'", code "'.$code.'" returned.');
             }
 
-            if ($zip->extractTo($to . $this->documentName, $this->getContentPath()) === false) {
+            if ($zip->extractTo($to.$this->documentName, $this->getContentPath()) === false) {
                 throw new InvalidArgumentException('Destination not reachable.');
             }
         }
@@ -77,7 +80,12 @@ abstract class Document implements DocumentInterface
      */
     public function getDocumentHash()
     {
-        return md5_file($this->documentPath);
+        $hash = md5_file($this->documentPath);
+        if ($hash === false) {
+            throw new HashException();
+        }
+
+        return $hash;
     }
 
     /**
@@ -98,31 +106,32 @@ abstract class Document implements DocumentInterface
 
     /**
      * @inherit
-     * @param \DOMDocument $template
      */
-    public abstract function cleanup(\DOMDocument $template);
+    abstract public function cleanup(\DOMDocument $template);
 
     /**
      * @inherit
      */
-    public abstract function getContentPath();
+    abstract public static function getContentPath();
 
     /**
      * @inherit
      */
-    public abstract function getNodePath();
+    abstract public function getNodePath();
 
     /**
      * @inherit
-     * @param int $type XPATH_* constant.
-     * @param bool $global Append global xpath //.
+     *
+     * @param int  $type   XPATH_* constant
+     * @param bool $global append global xpath //
      */
-    public abstract function getNodeName($type, $global = false);
+    abstract public function getNodeName($type, $global = false);
 
     /**
      * @inherit
-     * @param string $id Id as entered in placeholder.
-     * @param Tag $tag Container tag.
+     *
+     * @param string $id  id as entered in placeholder
+     * @param Tag    $tag container tag
      */
-    public abstract function getExpression($id, Tag $tag);
+    abstract public function getExpression(string $id, Tag $tag): ExtensionInterface;
 }
