@@ -3,6 +3,7 @@
 namespace PHPStamp\Document\WordDocument\Extension;
 
 use PHPStamp\Exception\ExtensionException;
+use PHPStamp\Exception\XmlException;
 use PHPStamp\Extension\Extension;
 use PHPStamp\Processor;
 use PHPStamp\XMLHelper;
@@ -11,11 +12,13 @@ class ListItem extends Extension
 {
     /**
      * @inherit
+     *
+     * @throws ExtensionException
      */
-    protected function prepareArguments(array $arguments)
+    protected function prepareArguments(array $arguments): array
     {
         if (count($arguments) !== 0) {
-            throw new ExtensionException('Wrong arguments number, 0 needed, got ' . count($arguments));
+            throw new ExtensionException('Wrong arguments number, 0 needed, got '.count($arguments));
         }
 
         return $arguments;
@@ -27,47 +30,72 @@ class ListItem extends Extension
     protected function insertTemplateLogic(array $arguments, \DOMElement $node)
     {
         $template = $node->ownerDocument;
+        if ($template === null) {
+            throw new XmlException('Detached node');
+        }
+
+        $root = $template->documentElement;
+        if ($root === null) {
+            throw new XmlException('Root node not found');
+        }
 
         $listName = $this->tag->getRelativePath();
+        if ($listName === null) {
+            throw new ExtensionException('Tag path is empty');
+        }
 
         // find existing or initiate new table row template
         if ($this->isListItemTemplateExist($listName, $template) === false) {
-
             $rowTemplate = $template->createElementNS(Processor::XSL_NS, 'xsl:template');
             $rowTemplate->setAttribute('name', $listName);
 
             // find row node
             $rowNode = XMLHelper::parentUntil('w:p', $node);
+            if ($rowNode === null) {
+                throw new ExtensionException('Cant find row node');
+            }
+
+            $containerNode = $rowNode->parentNode;
+            if ($containerNode === null) {
+                throw new ExtensionException('Cant find container node');
+            }
 
             // call-template for each row
             $foreachNode = $template->createElementNS(Processor::XSL_NS, 'xsl:for-each');
-            $foreachNode->setAttribute('select', '/' . Processor::VALUE_NODE . '/' . $listName . '/item');
+            $foreachNode->setAttribute('select', '/'.Processor::VALUE_NODE.'/'.$listName.'/item');
             $callTemplateNode = $template->createElementNS(Processor::XSL_NS, 'xsl:call-template');
             $callTemplateNode->setAttribute('name', $listName);
             $foreachNode->appendChild($callTemplateNode);
 
             // insert call-template before moving
-            $rowNode->parentNode->insertBefore($foreachNode, $rowNode);
+            $containerNode->insertBefore($foreachNode, $rowNode);
 
             // move node into row template
             $rowTemplate->appendChild($rowNode);
-            $template->documentElement->appendChild($rowTemplate);
-
+            $root->appendChild($rowTemplate);
         }
 
         // FIXME пофиксить повторное использование функции
         Processor::insertTemplateLogic($this->tag->getTextContent(), '.', $node);
     }
 
-    private function isListItemTemplateExist($rowName, \DOMDocument $template)
+    /**
+     * @throws ExtensionException
+     * @throws XmlException
+     */
+    private function isListItemTemplateExist(string $rowName, \DOMDocument $template): bool
     {
         $xpath = new \DOMXPath($template);
-        $nodeList = $xpath->query('/xsl:stylesheet/xsl:template[@name="' . $rowName . '"]');
+
+        $nodeList = $xpath->query('/xsl:stylesheet/xsl:template[@name="'.$rowName.'"]');
+        if ($nodeList === false) {
+            throw new XmlException('Malformed query');
+        }
 
         if ($nodeList->length > 1) {
             throw new ExtensionException('Unexpected template count.');
         }
 
-        return ($nodeList->length === 1);
+        return $nodeList->length === 1;
     }
-} 
+}
